feat: remove feedback button
Has been removed via a contribution to Jaen:

A lot of websites build using Jaen are targeted at non-engish speaking
users and the feedback form is in english. This is not ideal and
therefore the feedback form should be disabled by default.

This commit does not add any new functionality that makes it posible to
enable the feedback form again. This will be added in a future commit.